resume-matcher/src/parser.py
from docx import Document
import fitz  # PyMuPDF

def extract_text_from_pdf(file_path):
    doc = fitz.open(file_path)
    return "\n".join([page.get_text() for page in doc])

def extract_text_from_docx(file_path):
    doc = Document(file_path)
    return "\n".join([p.text for p in doc.paragraphs])

def load_job_description(file_path):
    with open(file_path, 'r') as f:
        return f.read()
