resume-matcher/requirements.txt
sentence-transformers
torch
PyMuPDF
python-docx
streamlit
